worked on login, register and some front end rearrangement

+++ app.py
@@ -1,12 +1,15 @@
-from flask import Flask, render_template, request, redirect, session
+from flask import Flask, render_template, request, redirect, session, flash
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime
+from helper import validate, number_or_none, date_or_none, strip_if_not_none, login_required
+from werkzeug.security import generate_password_hash, check_password_hash
+
 
 app = Flask(__name__)
 
 
 app.config['SQLALCHEMY_DATABASE_URI'] = "sqlite:///test.db"
-app.config['SECRET_KEY'] = 'super secret key'
+app.config['SECRET_KEY'] = '$catonthekeyboard'
 db = SQLAlchemy(app)
 
 
@@ -135,13 +138,13 @@ class Counselling(db.Model):
 
     def __repr__(self):
         return f"{self.record}"
-
-
-
+    
+    
 
 @app.route("/")
+@login_required
 def index():
-    return "hello world"
+    return redirect("/general")
 
 
 @app.route("/register", methods=['GET', 'POST'])
@@ -149,16 +152,23 @@ def register():
     if request.method == 'POST':
         email = request.form.get('email')
         password = request.form.get('password')
+        confirm = request.form.get('confirm')
         existing_user = Users.query.filter_by(email=email).first()
 
-        if existing_user is None:
-            new_user = Users(email=email, password=password)
-            db.session.add(new_user)
-            db.session.commit()
-            return render_template('login.html')
-        else:
-            # TODO
-            return 'user already exists'
+        if existing_user:
+            flash("user with that email already exists")
+            return redirect("/register")
+        
+        if password != confirm:
+            flash("The passwords don't match")
+            return redirect("/register")
+        
+        new_user = Users(email=email, password=generate_password_hash(password))
+        db.session.add(new_user)
+        db.session.commit()
+
+        flash("User succesfully registered!")
+        return redirect("/login")
     else:
         return render_template("register.html")
         
@@ -169,54 +179,64 @@ def login():
         email=request.form.get('email')
         password=request.form.get('password')
         user = Users.query.filter_by(email=email).first()
+
+        if not user:
+            flash("User doesn't exists")
+            return redirect("/login")
         
-        if user and user.password == password:
+        if check_password_hash(user.password, password):
             session["user_id"] = user.id
             return redirect("/general")
         else:
-            # TODO
-            return 'wrong password or user doesnt exists'
+            flash("wrong password")
+            return redirect("/login")
     else:
         return render_template("login.html")
         
 
+@app.route("/logout", methods=['POST'])
+@login_required
+def logout():
+    session.clear()
+    flash("logged out successfully")
+    return redirect("/login")
 
-@app.route("/general", methods=['GET', 'POST'])
+
+@app.route("/general", methods=['GET'])
+@login_required
 def general():
-    if request.method == 'POST':
-        # TODO
-        pass
-    else:
-        general_info = General.query.filter_by(user_id=session["user_id"]).first()
-        return render_template("general.html", general=general_info)
+    general_info = General.query.filter_by(user_id=session["user_id"]).first()
+    return render_template("general.html", general=general_info)
 
 
-@app.route("/semester<int:semester_number>", methods=['GET', 'POST'])
+@app.route("/semester<int:semester_number>", methods=['GET'])
+@login_required
 def semester(semester_number):
 
-    if request.method == 'POST':
-        pass
-    else:
-        attendance_records = Attendance.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
-        iat_records = Iat.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
-        event_records = Event.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
-        mooc_records = Mooc.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
-        project_records = Project.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
-        counselling_records = Counselling.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
-        
-        return render_template("semester.html", 
-                               semester_number=semester_number,
-                               attendance_records=attendance_records,
-                               iat_records=iat_records,
-                               event_records=event_records,
-                               mooc_records=mooc_records,
-                               project_records=project_records,
-                               counselling_records=counselling_records)
+    if not (1 <= semester_number <= 8):
+        return redirect("/general")
+
+    attendance_records = Attendance.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
+    iat_records = Iat.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
+    event_records = Event.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
+    mooc_records = Mooc.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
+    project_records = Project.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
+    counselling_records = Counselling.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
+    
+    return render_template("semester.html", 
+                            semester_number=semester_number,
+                            attendance_records=attendance_records,
+                            iat_records=iat_records,
+                            event_records=event_records,
+                            mooc_records=mooc_records,
+                            project_records=project_records,
+                            counselling_records=counselling_records)
     
 
 
 
 @app.route("/semester<int:semester_number>-edit", methods=['GET', 'POST'])
+@login_required
 def edit_semester(semester_number):
     if request.method == 'POST':
         attendance_rows = int(request.form.get('attendance-rows'))
@@ -225,6 +245,9 @@ def edit_semester(semester_number):
         mooc_rows = int(request.form.get('mooc-rows'))
         project_rows = int(request.form.get('project-rows'))
         counselling_rows = int(request.form.get('counselling-rows'))
+
+
+        print(f"{attendance_rows} {iat_rows} {event_rows} {mooc_rows} {project_rows} {counselling_rows}")
 
 
         prev_attendance = Attendance.query.filter_by(user_id=session['user_id'], semester=semester_number).all()
@@ -265,76 +288,96 @@ def edit_semester(semester_number):
 
         db.session.commit()
 
-        # for i in range(attendance_rows):
-        #     code = request.form.get(f"attendance-row{i + 1}-code")
-        #     subject = request.form.get(f"attendance-row{i + 1}-subject")
-        #     t1 = request.form.get(f"attendance-row{i + 1}-T1")
-        #     a1 = request.form.get(f"attendance-row{i + 1}-A1")
-        #     t2 = request.form.get(f"attendance-row{i + 1}-T2")
-        #     a2 = request.form.get(f"attendance-row{i + 1}-A2")
-        #     t3 = request.form.get(f"attendance-row{i + 1}-T3")
-        #     a3 = request.form.get(f"attendance-row{i + 1}-A3")
+        for i in range(attendance_rows):
+            code = strip_if_not_none(request.form.get(f"attendance-row{i + 1}-code"))
+            subject = strip_if_not_none(request.form.get(f"attendance-row{i + 1}-subject"))
+            t1 = number_or_none(strip_if_not_none(request.form.get(f"attendance-row{i + 1}-T1")))
+            a1 = number_or_none(strip_if_not_none(request.form.get(f"attendance-row{i + 1}-A1")))
+            t2 = number_or_none(strip_if_not_none(request.form.get(f"attendance-row{i + 1}-T2")))
+            a2 = number_or_none(strip_if_not_none(request.form.get(f"attendance-row{i + 1}-A2")))
+            t3 = number_or_none(strip_if_not_none(request.form.get(f"attendance-row{i + 1}-T3")))
+            a3 = number_or_none(strip_if_not_none(request.form.get(f"attendance-row{i + 1}-A3")))
 
-        #     record = Attendance(user_id=session['user_id'], code=code,
-        #                         subject=subject, t1=t1, a1=a1, t2=t2, a2=a2, 
-        #                         t3=t3, a3=a3, semester=semester_number)
+            if not code and not subject:
+                continue
+
+
+            record = Attendance(user_id=session['user_id'], code=code,
+                                subject=subject, t1=t1, a1=a1, t2=t2, a2=a2, 
+                                t3=t3, a3=a3, semester=semester_number)
             
-        #     db.session.add(record)
+            db.session.add(record)
 
 
 
-        # for i in range(iat_rows):
-        #     code = request.form.get(f"iat-row{i + 1}-code")
-        #     subject = request.form.get(f"iat-row{i + 1}-subject")
-        #     max = request.form.get(f"iat-row{i + 1}-max")
-        #     iat1 = request.form.get(f"iat-row{i + 1}-iat1")
-        #     iat2 = request.form.get(f"iat-row{i + 1}-iat2")
-        #     iat3 = request.form.get(f"iat-row{i + 1}-iat3")
+        for i in range(iat_rows):
+            code = strip_if_not_none(request.form.get(f"iat-row{i + 1}-code"))
+            subject = strip_if_not_none(request.form.get(f"iat-row{i + 1}-subject"))
+            max = number_or_none(strip_if_not_none(request.form.get(f"iat-row{i + 1}-max")))
+            iat1 = number_or_none(strip_if_not_none(request.form.get(f"iat-row{i + 1}-iat1")))
+            iat2 = number_or_none(strip_if_not_none(request.form.get(f"iat-row{i + 1}-iat2")))
+            iat3 = number_or_none(strip_if_not_none(request.form.get(f"iat-row{i + 1}-iat3")))
 
-        #     record = Iat(user_id=session['user_id'], code=code, subject=subject,
-        #                  max=max, iat1=iat1, iat2=iat2, iat3=iat3, semester=semester_number)
+            if not code and not subject:
+                continue
+
+            record = Iat(user_id=session['user_id'], code=code, subject=subject,
+                         max=max, iat1=iat1, iat2=iat2, iat3=iat3, semester=semester_number)
             
-        #     db.session.add(record)
+            db.session.add(record)
 
 
-        # for i in range(event_rows):
-        #     name = request.form.get(f"events-row{i + 1}-name")
-        #     title = request.form.get(f"events-row{i + 1}-title")
-        #     date = request.form.get(f"events-row{i + 1}-date")
+        for i in range(event_rows):
+            name = strip_if_not_none(request.form.get(f"events-row{i + 1}-name"))
+            title = strip_if_not_none(request.form.get(f"events-row{i + 1}-title"))
+            date = date_or_none(strip_if_not_none(request.form.get(f"events-row{i + 1}-date")))
 
-        #     record = Event(user_id=session['user_id'], club_name=name, 
-        #                    event_title=title, event_date=date, semester=semester_number)
+            if not name and not title:
+                continue
+
+            record = Event(user_id=session['user_id'], club_name=name, 
+                           event_title=title, event_date=date, semester=semester_number)
             
-        #     db.session.add(record)
+            db.session.add(record)
 
-        # for i in range(mooc_rows):
-        #     under = request.form.get(f"mooc-row{i + 1}-under")
-        #     title = request.form.get(f"mooc-row{i + 1}-title")
-        #     start = request.form.get(f"mooc-row{i + 1}-start")
-        #     completed = request.form.get(f"mooc-row{i + 1}-completed")
-        #     score = request.form.get(f"mooc-row{i + 1}-score")
+        for i in range(mooc_rows):
+            under = strip_if_not_none(request.form.get(f"mooc-row{i + 1}-under"))
+            title = strip_if_not_none(request.form.get(f"mooc-row{i + 1}-title"))
+            start = date_or_none(strip_if_not_none(request.form.get(f"mooc-row{i + 1}-start")))
+            completed = date_or_none(strip_if_not_none(request.form.get(f"mooc-row{i + 1}-completed")))
+            score = number_or_none(strip_if_not_none(request.form.get(f"mooc-row{i + 1}-score")))
 
-        #     record = Mooc(user_id=session['user_id'], under=under, title=title, 
-        #                   start_date=start, completed_date=completed,
-        #                     score=score, semester=semester_number)
+            if not title:
+                continue
 
-        #     db.session.add(record)
+            record = Mooc(user_id=session['user_id'], under=under, title=title, 
+                          start_date=start, completed_date=completed,
+                            score=score, semester=semester_number)
+
+            db.session.add(record)
 
 
-        # for i in range(project_rows):
-        #     title = request.form.get(f"project-row{i + 1}-title")
-        #     hours = request.form.get(f"project-row{i + 1}-hours")
-        #     start = request.form.get(f"project-row{i + 1}-start")
-        #     completed = request.form.get(f"project-row{i + 1}-completed")
+        for i in range(project_rows):
+            title = strip_if_not_none(request.form.get(f"project-row{i + 1}-title"))
+            hours = number_or_none(strip_if_not_none(request.form.get(f"project-row{i + 1}-hours")))
+            start = date_or_none(strip_if_not_none(request.form.get(f"project-row{i + 1}-start")))
+            completed = date_or_none(strip_if_not_none(request.form.get(f"project-row{i + 1}-completed")))
 
-        #     record = Project(user_id=session['user_id'], title=title, man_hours=hours,
-        #                      start_date=start, completed_date=completed, semester=semester_number)
+            if not title:
+                continue
+
+            record = Project(user_id=session['user_id'], title=title, man_hours=hours,
+                             start_date=start, completed_date=completed, semester=semester_number)
             
-        #     db.session.add(record)
+            db.session.add(record)
+
 
         for i in range(counselling_rows):
-            date = datetime.strptime(request.form.get(f"counselling-row{i + 1}-date"), "%Y-%m-%d").date()
-            record = request.form.get(f"counselling-row{i + 1}-record")
+            date = date_or_none(strip_if_not_none(request.form.get(f"counselling-row{i + 1}-date")))
+            record = strip_if_not_none(request.form.get(f"counselling-row{i + 1}-record"))
+
+            if not record:
+                continue
 
             row = Counselling(user_id=session['user_id'], date=date, record=record, semester=semester_number)
 
@@ -358,7 +401,8 @@ def edit_semester(semester_number):
                                event_records=event_records,
                                mooc_records=mooc_records,
                                project_records=project_records,
-                               counselling_records=counselling_records)
+                               counselling_records=counselling_records,
+                               semester_number=semester_number)
 
 
 
@@ -366,6 +410,7 @@ def edit_semester(semester_number):
 
 
 @app.route("/general-edit", methods=['GET', 'POST'])
+@login_required
 def edit_general():
     if request.method == 'POST':
         name = request.form.get('name')
@@ -393,7 +438,8 @@ def edit_general():
         
         #delete existing record
         to_be_deleted = General.query.filter_by(user_id=session['user_id']).first()
-        db.session.delete(to_be_deleted)
+        if to_be_deleted:
+            db.session.delete(to_be_deleted)
 
         #add new record
         db.session.add(general_info)
